refactor: move components to components folder
