+++ src/app/core/tndm-ai-agent/components/tndm-ai-agent-button/tndm-ai-agent-button.scss
@@ -1,5 +1,5 @@
-@use '../../../../styles/variables' as vars;
-@use '../../../../styles/typography' as typ;
+@use '../../../../../styles/variables' as vars;
+@use '../../../../../styles/typography' as typ;
 
 :host {
   position: absolute;
