+++ src/app/core/tndm-ai-agent/components/tndm-ai-input/tndm-ai-input.component.scss
@@ -1,5 +1,5 @@
-@use '../../../../styles/typography' as typ;
-@use '../../../../styles/variables' as vars;
+@use '../../../../../styles/typography' as typ;
+@use '../../../../../styles/variables' as vars;
 
 :host {
   display: block;
